requirements.txt
asgiref==3.8.1
Django==5.0.4
django-taggit==5.0.1
django-widget-tweaks==1.5.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2024.1
